services/push_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pika==1.3.2
requests==2.31.0
pydantic==2.5.0
redis==5.0.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
